projects/casting-agency/requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ecdsa==0.19.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
